Struct roughenough::RtMessage

pub struct RtMessage { /* fields omitted */ }

A Roughtime protocol message; a map of u32 tags to arbitrary byte-strings.

Methods

impl RtMessage

pub fn new(num_fields: u32) -> Self

Construct a new RtMessage

Arguments

num_fields - Reserve space for this many fields.

pub fn from_bytes(bytes: &[u8]) -> Result<Self, Error>

Construct a new RtMessage from the on-the-wire representation in bytes

Arguments

bytes - On-the-wire representation

pub fn new_deliberately_invalid(tags: Vec<Tag>, values: Vec<Vec<u8>>) -> Self

Dangerous: construct a new RtMessage without validation or error checking.

Intended only for construction of deliberately bogus responses as part of Roughtime's ecosystem.

pub fn add_field(&mut self, tag: Tag, value: &[u8]) -> Result<(), Error>

Add a field to this RtMessage

Arguments

tag - The Tag to add. Tags must be added in strictly increasing order, violating this will result in a Error::TagNotStrictlyIncreasing.

value - Value for the tag.

pub fn get_field(&self, tag: Tag) -> Option<&[u8]>

Retrieve the value associated with tag, if present.

Arguments

tag - The Tag to try and retrieve.

pub fn num_fields(&self) -> u32

Returns the number of tag/value pairs in the message

pub fn tags(&self) -> &[Tag]

Returns a slice of the tags in the message

pub fn values(&self) -> &[Vec<u8>]

Returns a slice of the values in the message

pub fn into_hash_map(self) -> HashMap<Tag, Vec<u8>>

Converts the message into a HashMap mapping each tag to its value

pub fn encode(&self) -> Result<Vec<u8>, Error>

Encode this message into its on-the-wire representation.

pub fn encoded_size(&self) -> usize

Returns the length in bytes of this message's on-the-wire representation.

pub fn pad_to_kilobyte(&mut self)

Adds a PAD tag to the end of this message, with a length set such that the final encoded size of this message is 1KB

If the encoded size of this message is already >= 1KB, this method does nothing